+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-08-06
+
 ### Added
 
 - The original colors are now restored automatically on the next startup after quitting VS Code with gaming mode running
@@ -38,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/akiomik/vscode-gaming/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/akiomik/vscode-gaming/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/akiomik/vscode-gaming/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/akiomik/vscode-gaming/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/akiomik/vscode-gaming/releases/tag/v0.0.1